test-fbc: add the files needed to build a test catalog
This commit adds a new catalog for trustee-operator, that contains
only the latest bundle reference.
We will create a nudge relationship between the bundle image and this test-fbc
component in Konflux, so that everytime a new bundle is built, this unique
reference is updated.
The Dockerfile for this test-fbc then runs our render.sh script, looking for the
related images for the referenced bundle, and generating the catalog.json file
at build time.

This means that merging the nudge PR will automatically produce a catalog that
can be used on all versions of OCP.
It makes it easier to create a testable catalog for new builds.
Using specific tags on the catalog will also help identify what builds were
tested whenever a bug is reported.
Hopefully, this standardization should help with test automation.

Note that as the build of this catalog requires network access, it makes this
catalog unreleasable, as we can't build it hermetically.

Signed-off-by: Julien Ropé <jrope@redhat.com>